feat: search only features when there is search string
